requirements.txt
requests>=2.31.0
Pillow>=10.0.0
python-dotenv>=1.0.0
langchain>=0.2.0
langchain-core>=0.2.0
langchain-google-genai>=1.0.0
langchain-community>=0.2.0
langchain-ollama>=0.1.0
langgraph>=0.1.0
langgraph-checkpoint-sqlite>=1.0.0
fastapi>=0.100.0
uvicorn>=0.22.0
websockets>=11.0.3
pywebview>=4.0.0